+++ cargo-crev/src/review.rs
@@ -2,7 +2,7 @@ use crate::{
     edit, opts,
     opts::{CargoOpts, CrateSelector},
     prelude::*,
-    term,
+    term, url_to_status_str,
 };
 use anyhow::format_err;
 use crev_data::{
@@ -10,7 +10,7 @@ use crev_data::{
     Rating,
 };
 use crev_lib::{self, local::Local, TrustProofType};
-use std::default::Default;
+use std::{default::Default, fmt::Write};
 
 use crate::{repo::*, shared::*};
 
@@ -26,6 +26,7 @@ pub fn create_review_proof(
     proof_create_opt: &opts::CommonProofCreate,
     diff_version: &Option<Option<Version>>,
     skip_activity_check: bool,
+    show_override_suggestions: bool,
     cargo_opts: CargoOpts,
 ) -> Result<()> {
     let repo = Repo::auto_open_cwd(cargo_opts)?;
@@ -104,48 +105,59 @@ pub fn create_review_proof(
     let id = local.read_current_unlocked_id(&term::read_passphrase)?;
 
     let db = local.load_db()?;
-    let mut review = proof::review::PackageBuilder::default()
-        .from(id.id.to_owned())
-        .package(proof::PackageInfo {
-            id: proof::PackageVersionId::new(
-                PROJECT_SOURCE_CRATES_IO.to_owned(),
-                crate_.name().to_string(),
-                effective_crate_version.to_owned(),
-            ),
-            digest: digest_clean.into_vec(),
-            digest_type: proof::default_digest_type(),
-            revision: vcs_info_to_revision_string(vcs),
-            revision_type: proof::default_revision_type(),
-        })
-        .review(if advise_common.is_some() || report_severity.is_some() {
-            crev_data::Review::new_none()
-        } else {
-            trust.to_review()
-        })
-        .diff_base(diff_base)
-        .build()
-        .map_err(|e| format_err!("{}", e))?;
 
-    let previous_date =
-        if let Some((prev_date, prev_review, prev_advisories, prev_issues, prev_comment)) =
-            find_previous_review_data(
-                &db,
-                &id.id,
+    let default_review_content = if advise_common.is_some() || report_severity.is_some() {
+        crev_data::Review::new_none()
+    } else {
+        trust.to_review()
+    };
+
+    let (previous_date, mut review) = if let Some(mut previous_review) = db
+        .get_pkg_review(
+            PROJECT_SOURCE_CRATES_IO,
+            &crate_.name(),
+            effective_crate_version,
+            &id.id.id,
+        )
+        .cloned()
+    {
+        if trust == TrustProofType::Untrust {
+            *previous_review.review_possibly_none_mut() = default_review_content;
+        }
+        (Some(previous_review.common.date), previous_review)
+    } else {
+        let mut fresh_review = proof::review::PackageBuilder::default()
+            .from(id.id.to_owned())
+            .package(proof::PackageInfo {
+                id: proof::PackageVersionId::new(
+                    PROJECT_SOURCE_CRATES_IO.to_owned(),
+                    crate_.name().to_string(),
+                    effective_crate_version.to_owned(),
+                ),
+                digest: digest_clean.into_vec(),
+                digest_type: proof::default_digest_type(),
+                revision: vcs_info_to_revision_string(vcs),
+                revision_type: proof::default_revision_type(),
+            })
+            .review(default_review_content)
+            .diff_base(diff_base)
+            .build()
+            .map_err(|e| format_err!("{}", e))?;
+
+        if let Some(diff_base_version) = diff_base_version.clone() {
+            if let Some(base_review) = db.get_pkg_review(
+                PROJECT_SOURCE_CRATES_IO,
                 &crate_.name(),
-                effective_crate_version,
                 &diff_base_version,
-            )
-        {
-            if trust != TrustProofType::Untrust {
-                *review.review_possibly_none_mut() = prev_review;
+                &id.id.id,
+            ) {
+                fresh_review.comment = base_review.comment.to_owned();
+                *fresh_review.review_possibly_none_mut() =
+                    base_review.review_possibly_none().to_owned()
             }
-            review.comment = prev_comment;
-            review.advisories = prev_advisories;
-            review.issues = prev_issues;
-            prev_date
-        } else {
-            None
-        };
+        }
+        (None, fresh_review)
+    };
 
     if let Some(advise_common) = advise_common {
         let mut advisory: proof::review::package::Advisory = advise_common.affected.into();
@@ -166,13 +178,36 @@ pub fn create_review_proof(
 
     review.alternatives = db.get_pkg_alternatives_by_author(&id.id.id, &review.package.id.id);
 
-    let review = edit::edit_proof_content_iteractively(
+    let mut review = edit::edit_proof_content_iteractively(
         &review,
         previous_date.as_ref(),
         diff_base_version.as_ref(),
         None,
+        |text| {
+            if show_override_suggestions && review.override_.is_empty() {
+                writeln!(text, "# override:")?;
+            }
+
+            if show_override_suggestions {
+                for review in db.get_package_reviews_for_package(
+                    PROJECT_SOURCE_CRATES_IO,
+                    Some(&pkg_id.name()),
+                    Some(&pkg_id.version()),
+                ) {
+                    let id = &review.common.from.id;
+                    let (status, url) = url_to_status_str(&db.lookup_url(id));
+                    writeln!(text, "# - id-type: {}", "crev")?; // TODO: support other ids?
+                    writeln!(text, "#   id: {}", id)?;
+                    writeln!(text, "#   url: {} # {}", url, status)?;
+                    writeln!(text, "#   comment: \"\"")?;
+                }
+            }
+
+            Ok(())
+        },
     )?;
 
+    review.touch_date();
     let proof = review.sign_by(&id)?;
 
     let commit_msg = format!(
@@ -181,45 +216,6 @@ pub fn create_review_proof(
         version = effective_crate_version
     );
     maybe_store(&local, &proof, &commit_msg, proof_create_opt)
-}
-
-pub fn find_previous_review_data(
-    db: &crev_wot::ProofDB,
-    id: &crev_data::PublicId,
-    name: &str,
-    crate_version: &Version,
-    diff_base_version: &Option<Version>,
-) -> Option<(
-    Option<crev_data::proof::Date>,
-    crev_data::proof::review::Review,
-    Vec<crev_data::proof::review::package::Advisory>,
-    Vec<crev_data::proof::review::package::Issue>,
-    String,
-)> {
-    if let Some(previous_review) =
-        db.get_pkg_review(PROJECT_SOURCE_CRATES_IO, name, crate_version, &id.id)
-    {
-        return Some((
-            Some(previous_review.common.date),
-            previous_review.review_possibly_none().to_owned(),
-            previous_review.advisories.to_owned(),
-            previous_review.issues.to_owned(),
-            previous_review.comment.to_owned(),
-        ));
-    } else if let Some(diff_base_version) = diff_base_version {
-        if let Some(base_review) =
-            db.get_pkg_review(PROJECT_SOURCE_CRATES_IO, name, diff_base_version, &id.id)
-        {
-            return Some((
-                None,
-                base_review.review_possibly_none().to_owned(),
-                vec![],
-                vec![],
-                base_review.comment.to_owned(),
-            ));
-        }
-    }
-    None
 }
 
 pub fn find_reviews(crate_: &opts::CrateSelector) -> Result<Vec<proof::review::Package>> {