+++ crev-lib/src/tests.rs
@@ -1,9 +1,9 @@
 use super::*;
 use crev_data::{
-    proof::{self, trust::TrustLevel, ContentExt},
+    proof::{self, trust::TrustLevel, ContentExt, OverrideItem},
     Digest, Level, UnlockedId, Url, Version,
 };
-use crev_wot::{FetchSource, ProofDB};
+use crev_wot::{FetchSource, ProofDB, TrustSet};
 use default::default;
 use std::{str::FromStr, sync::Arc};
 
@@ -57,8 +57,11 @@ pass:
     let locked = id::LockedId::from_str(yaml)?;
     let unlocked = locked.to_unlocked("a")?;
 
-    let _trust_proof =
-        unlocked.create_signed_trust_proof(vec![unlocked.as_public_id()], TrustLevel::High)?;
+    let _trust_proof = unlocked.create_signed_trust_proof(
+        vec![unlocked.as_public_id()],
+        TrustLevel::High,
+        vec![],
+    )?;
 
     Ok(())
 }
@@ -115,10 +118,10 @@ fn proofdb_distance() -> Result<()> {
         max_distance: 111,
     };
 
-    let a_to_b = a.create_signed_trust_proof(vec![b.as_public_id()], TrustLevel::High)?;
-    let b_to_c = b.create_signed_trust_proof(vec![c.as_public_id()], TrustLevel::Medium)?;
-    let c_to_d = c.create_signed_trust_proof(vec![d.as_public_id()], TrustLevel::Low)?;
-    let d_to_e = d.create_signed_trust_proof(vec![e.as_public_id()], TrustLevel::High)?;
+    let a_to_b = a.create_signed_trust_proof(vec![b.as_public_id()], TrustLevel::High, vec![])?;
+    let b_to_c = b.create_signed_trust_proof(vec![c.as_public_id()], TrustLevel::Medium, vec![])?;
+    let c_to_d = c.create_signed_trust_proof(vec![d.as_public_id()], TrustLevel::Low, vec![])?;
+    let d_to_e = d.create_signed_trust_proof(vec![e.as_public_id()], TrustLevel::High, vec![])?;
 
     let mut trustdb = ProofDB::new();
 
@@ -140,7 +143,7 @@ fn proofdb_distance() -> Result<()> {
     assert!(trust_set.contains(d.as_ref()));
     assert!(!trust_set.contains(e.as_ref()));
 
-    let b_to_d = b.create_signed_trust_proof(vec![d.as_public_id()], TrustLevel::Medium)?;
+    let b_to_d = b.create_signed_trust_proof(vec![d.as_public_id()], TrustLevel::Medium, vec![])?;
 
     trustdb.import_from_iter(vec![(b_to_d, url)].into_iter());
 
@@ -181,7 +184,7 @@ fn overwritting_reviews() -> Result<()> {
 
     let proof1 = a
         .as_public_id()
-        .create_package_review_proof(package.clone(), default(), "a".into())?
+        .create_package_review_proof(package.clone(), default(), vec![], "a".into())?
         .sign_by(&a)?;
     // it's lame, but oh well... ; we need to make sure there's a time delay between
     // the two proofs
@@ -189,7 +192,7 @@ fn overwritting_reviews() -> Result<()> {
     std::thread::sleep_ms(1);
     let proof2 = a
         .as_public_id()
-        .create_package_review_proof(package.clone(), default(), "b".into())?
+        .create_package_review_proof(package.clone(), default(), vec![], "b".into())?
         .sign_by(&a)?;
 
     for order in vec![vec![proof1.clone(), proof2.clone()], vec![proof2, proof1]] {
@@ -254,7 +257,7 @@ fn dont_consider_an_empty_review_as_valid() -> Result<()> {
 
     let proof1 = a
         .as_public_id()
-        .create_package_review_proof(package, review, "a".into())?
+        .create_package_review_proof(package, review, vec![], "a".into())?
         .sign_by(&a)?;
 
     let mut trustdb = ProofDB::new();
@@ -293,11 +296,15 @@ fn proofdb_distrust() -> Result<()> {
         max_distance: 10000,
     };
 
-    let a_to_bc =
-        a.create_signed_trust_proof(vec![b.as_public_id(), c.as_public_id()], TrustLevel::High)?;
-    let b_to_d = b.create_signed_trust_proof(vec![d.as_public_id()], TrustLevel::Low)?;
-    let d_to_c = d.create_signed_trust_proof(vec![c.as_public_id()], TrustLevel::Distrust)?;
-    let c_to_e = c.create_signed_trust_proof(vec![e.as_public_id()], TrustLevel::Low)?;
+    let a_to_bc = a.create_signed_trust_proof(
+        vec![b.as_public_id(), c.as_public_id()],
+        TrustLevel::High,
+        vec![],
+    )?;
+    let b_to_d = b.create_signed_trust_proof(vec![d.as_public_id()], TrustLevel::Low, vec![])?;
+    let d_to_c =
+        d.create_signed_trust_proof(vec![c.as_public_id()], TrustLevel::Distrust, vec![])?;
+    let c_to_e = c.create_signed_trust_proof(vec![e.as_public_id()], TrustLevel::Low, vec![])?;
 
     let mut trustdb = ProofDB::new();
 
@@ -321,7 +328,8 @@ fn proofdb_distrust() -> Result<()> {
 
     // This introduces a tie between nodes banning each other.
     // Both should be removed from the trust_set.
-    let e_to_d = e.create_signed_trust_proof(vec![d.as_public_id()], TrustLevel::Distrust)?;
+    let e_to_d =
+        e.create_signed_trust_proof(vec![d.as_public_id()], TrustLevel::Distrust, vec![])?;
 
     trustdb.import_from_iter(vec![(e_to_d, url)].into_iter());
 
@@ -337,5 +345,83 @@ fn proofdb_distrust() -> Result<()> {
     assert!(!trust_set.contains(d.as_ref()));
     assert!(!trust_set.contains(e.as_ref()));
 
+    Ok(())
+}
+
+#[test]
+fn proofdb_trust_ignore_override() -> Result<()> {
+    let url = FetchSource::Url(Arc::new(Url::new_git("https://a")));
+    let a = UnlockedId::generate_for_git_url("https://a");
+    let b = UnlockedId::generate_for_git_url("https://b");
+    let c = UnlockedId::generate_for_git_url("https://c");
+    let d = UnlockedId::generate_for_git_url("https://d");
+
+    let distance_params = TrustDistanceParams {
+        high_trust_distance: 1,
+        medium_trust_distance: 10,
+        low_trust_distance: 100,
+        max_distance: 10000,
+    };
+
+    // a trust b and c, but c more, c overrides (ignores) trust of b in d
+    let a_to_b = a.create_signed_trust_proof(vec![b.as_public_id()], TrustLevel::Medium, vec![])?;
+    let a_to_c = a.create_signed_trust_proof(vec![c.as_public_id()], TrustLevel::High, vec![])?;
+
+    let b_to_d = b.create_signed_trust_proof(vec![d.as_public_id()], TrustLevel::High, vec![])?;
+
+    let c_to_d = {
+        let mut c_to_d_unsigned =
+            c.id.create_trust_proof(vec![d.as_public_id()], TrustLevel::None, vec![])?;
+        c_to_d_unsigned.override_.push(OverrideItem {
+            id: b.as_public_id().clone(),
+            comment: "".into(),
+        });
+        c_to_d_unsigned.sign_by(&c)?
+    };
+
+    let mut trustdb = ProofDB::new();
+
+    trustdb.import_from_iter(
+        vec![a_to_b, a_to_c.clone(), b_to_d.clone(), c_to_d.clone()]
+            .into_iter()
+            .map(|x| (x, url.clone())),
+    );
+
+    let trust_set: TrustSet = trustdb.calculate_trust_set(a.as_ref(), &distance_params);
+    let trusted_ids: HashSet<_> = trust_set.trusted_ids().cloned().collect();
+
+    assert!(trust_set
+        .trust_ignore_overrides
+        .contains_key(&(b.id.id.clone(), d.id.id.clone())));
+    assert!(trusted_ids.contains(a.as_ref()));
+    assert!(trusted_ids.contains(b.as_ref()));
+    assert!(trusted_ids.contains(c.as_ref()));
+    assert!(!trusted_ids.contains(d.as_ref()));
+
+    // had the `a` trusted `b` at the same level as `c`, the override to ignore `b`'s trust in `d`
+    // would not have effect from PoV of `a`
+    {
+        let a_to_b =
+            a.create_signed_trust_proof(vec![b.as_public_id()], TrustLevel::High, vec![])?;
+
+        let mut trustdb = ProofDB::new();
+
+        trustdb.import_from_iter(
+            vec![a_to_b, a_to_c, b_to_d, c_to_d]
+                .into_iter()
+                .map(|x| (x, url.clone())),
+        );
+
+        let trust_set: TrustSet = trustdb.calculate_trust_set(a.as_ref(), &distance_params);
+        let trusted_ids: HashSet<_> = trust_set.trusted_ids().cloned().collect();
+
+        assert!(trust_set
+            .trust_ignore_overrides
+            .contains_key(&(b.id.id.clone(), d.id.id.clone())));
+        assert!(trusted_ids.contains(a.as_ref()));
+        assert!(trusted_ids.contains(b.as_ref()));
+        assert!(trusted_ids.contains(c.as_ref()));
+        assert!(trusted_ids.contains(d.as_ref()));
+    }
     Ok(())
 }